requirements.txt
streamlit
tensorflow-cpu
torch
torchvision
ultralytics
opencv-python-headless
numpy
pandas
matplotlib
seaborn
plotly
scikit-learn
Pillow